requirements.txt
argparse
termcolor
azure-identity>=1.5.0
azure-mgmt-automation